refactor: Update FileListCommand.php to handle file availability
The `FileListCommand.php` file was modified to handle file availability. The `available` property was added to the `FileListCommand` class and set to `true` by default. When an error occurs during file listing, the `available` property is set to `false`. This change ensures that the `FileListCommandScout` class only retrieves files if the `available` property is `true`.