refactor: Add skipFilenames method to FileList class
The FileList class was modified to include a new method called skipFilenames. This method allows users to specify a list of filenames to skip when scanning files. The skipFilenames method takes an array of filenames as input and updates the skipFilenames property accordingly. This change provides more flexibility in excluding specific files from the file scanning process.