chore: Add `getArguments` and `toArray` methods to FileListCommand class
The `getArguments` method was added to the FileListCommand class to retrieve the command arguments. This change enhances the functionality of the class by providing access to the arguments, which can be useful for customization or debugging purposes.

The `toArray` method was also added to the FileListCommand class. This method returns an array representation of the FileListCommand object, including its name, command, user, output, outputArray, success, noMemoryLimit, throwOnError, available, and errors properties. This change improves the usability of the class by allowing easy conversion to an array format.

These additions improve the FileListCommand class and provide more flexibility and convenience for working with command-related information.