feat: Add onlyExtensions method to FileList class
This commit adds the `onlyExtensions` method to the `FileList` class. The `onlyExtensions` method allows users to specify a list of file extensions to include when scanning for files. This enhances the functionality of the class by providing more control over the types of files to be included in the file list.